fix page register
