+++ src/pages/register/index.js
@@ -1,6 +1,5 @@
-import { createRegister } from '../../services/index.js';
-import { googleLogin } from '../../services/index.js';
-import { redirectTologinOrRegister } from '../../services/index.js';
+import { createRegister, googleLogin } from '../../services/index.js';
+import { redirectTologinOrRegister } from '../../router.js';
 
 export const Register = () => {
